backend/requirements.txt
scikit-learn==1.4.0
umap-learn==0.5.5
pandas==2.2.0
numpy==1.26.3 